# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación API ",
        "description": "Documentación API ",
        "contact": {
            "name": "API Support",
            "url": "http://www.fmt.com.do/support",
            "email": "info@fmt.com.do"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "description": "Listado de Usuarios",
                "operationId": "usuarios_usuarios_index",
                "parameters": [
                    {
                        "name": "activo",
                        "in": "query",
                        "description": "Si todos los usuarios o solo los activos, si pasan cualquier valor solo retorna los usuarios activos",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": "true"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/usuarios/store": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "description": "Crear un nuevo usuario",
                "operationId": "usuarios_usuarios_store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "role",
                                    "institution_id",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "description": "Email del usuario",
                                        "type": "string",
                                        "example": "user@email.aqui"
                                    },
                                    "role": {
                                        "description": "Role ID a asignar",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "institution_id": {
                                        "description": "Institution ID a asignar",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "telefono": {
                                        "description": "Telefono Contacto",
                                        "type": "string",
                                        "example": "(999) 555-5555"
                                    },
                                    "movil": {
                                        "description": "Movil Contacto",
                                        "type": "string",
                                        "example": "(999) 444-3333"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "SuperStrongPasswordHere"
                                    },
                                    "password_confirmation": {
                                        "description": "Password Confirmation",
                                        "type": "string",
                                        "example": "SuperStrongPasswordHere"
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/usuarios/{usuario}/show": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "description": "Mostrar un usuario",
                "operationId": "usuarios_usuarios_show",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "description": "Id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/usuarios/{usuario}/update": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "description": "Actualizar un usuario",
                "operationId": "usuarios_usuarios_update",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "description": "Id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "role",
                                    "institution_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "description": "Email del usuario",
                                        "type": "string",
                                        "example": "user@email.aqui"
                                    },
                                    "role": {
                                        "description": "Role ID Asignado",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "institution_id": {
                                        "description": "Institution ID Asignado",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "telefono": {
                                        "description": "Telefono Contacto",
                                        "type": "string",
                                        "example": "(999) 555-5555"
                                    },
                                    "movil": {
                                        "description": "Movil Contacto",
                                        "type": "string",
                                        "example": "(999) 444-3333"
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "Json",
                                            "example": "[...]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}